+++ CHANGELOG.md
@@ -22,6 +22,7 @@ All notable changes to this project will be documented in this file.
 lists they are listed as multiple entries.
 - Add new widget `ScanStatus` that indicates if the MPD database is being updated.
 - Add new global keybinds for Update and Rescan actions.
+- addrandom CLI command and `AddRandom` action which displays a modal allowing you to add random songs to the queue
 
 ### Changed